fix: Save correct version of the bikeshare.py file
The initial code was an html code although the extension of the file
was .py. This happened when the Python file was downloaded from the
Python project workspace. The downloaded file was actually an exe file.
This was fixed by a simple copy-paste from the Python code on the
project workspace to the bikeshare.py file.