fix: Save correct version of the bikeshare.py before refactoring
There was an error with the initial bikeshare.py file moved to the
local repository. The error occurred when the Python file was downloaded
from the Python project workspace. The downloaded file was actually an
exe file. This was fixed by a simple copy-paste from the Python code on
the project workspace to the bikeshare.py file.